Thinned subset of posterior sample from a Bayesian analysis of perception of correlation.

Data from Kay and Heer (2016), primarily used for testing and examples.

Details

For more details, see Kay and Heer (2016) or the Github repository describing the analysis: https://github.com/mjskay/ranking-correlation. The original experiment (but not this analysis of it) is described in Harrison et al. (2014).

This version of the posterior sample is thinned substantially and has omitted several parameters in order for it to be a more manageable size.

References

Kay, Matthew, and Jeffrey Heer. "Beyond Weber's law: A second look at ranking visualizations of correlation." IEEE transactions on visualization and computer graphics 22.1 (2016): 469-478.

Harrison, Lane, Fumeng Yang, Steven Franconeri, and Remco Chang. "Ranking visualizations of correlation using Weber's law." IEEE transactions on visualization and computer graphics 20.12 (2014): 1943-1952.