Density bins as data frames suitable for use with predict_curve

Generates a data frame of bins representing the kernel density (or histogram) of a vector, suitable for use in generating predictive distributions using predict_curve.

density_bins(x, n = 101, ...)

histogram_bins(x, n = 30, breaks = n, ...)

Arguments

x
A numeric vector

n
Number of bins

...
Additional arguments passed to density or hist.

breaks
Used to set bins for histogram_bins. Can be number of bins (by default it is set to the value of n) or a method for setting bins. See the breaks argument of hist.

Value

A data frame representing bins and their densities with the following columns:

mid

Bin midpoint

lower

Lower endpoint of each bin

upper

Upper endpoint of each bin

density

Density estimate of the bin

Details

These functions are simple wrappers to density and hist that compute density estimates and return their results in a consistent format: a data frame of bins suitable for use with predict_curve.

density_bins computes a kernel density estimate using density.

histogram_bins computes a density histogram using hist.

See also

Examples

library(ggplot2)
library(dplyr)
library(rstanarm)
library(modelr)

theme_set(theme_light())

m_mpg = stan_glm(mpg ~ hp * cyl, data = mtcars,
  # 1 chain / few iterations just so example runs quickly
  # do not use in practice
  chains = 1, iter = 500)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 500 [  0%]  (Warmup)
#> Iteration:  50 / 500 [ 10%]  (Warmup)
#> Iteration: 100 / 500 [ 20%]  (Warmup)
#> Iteration: 150 / 500 [ 30%]  (Warmup)
#> Iteration: 200 / 500 [ 40%]  (Warmup)
#> Iteration: 250 / 500 [ 50%]  (Warmup)
#> Iteration: 251 / 500 [ 50%]  (Sampling)
#> Iteration: 300 / 500 [ 60%]  (Sampling)
#> Iteration: 350 / 500 [ 70%]  (Sampling)
#> Iteration: 400 / 500 [ 80%]  (Sampling)
#> Iteration: 450 / 500 [ 90%]  (Sampling)
#> Iteration: 500 / 500 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.081 seconds (Warm-up)
#>                0.07 seconds (Sampling)
#>                0.151 seconds (Total)
#> 

step = 1
mtcars %>%
  group_by(cyl) %>%
  data_grid(hp = seq_range(hp, by = step)) %>%
  add_predicted_draws(m_mpg) %>%
  do(density_bins(.$pred)) %>%
  ggplot() +
  geom_rect(aes(
    xmin = hp - step/2, ymin = lower, ymax = upper, xmax = hp + step/2,
    fill = ordered(cyl), alpha = density
  )) +
  geom_point(aes(x = hp, y = mpg, color = ordered(cyl)), data = mtcars) +
  scale_alpha_continuous(range = c(0, 0.75))

#> Warning: Unknown or uninitialised column: 'pred'.

#> Error in density.default(x, n = n, ...): argument 'x' must be numeric