Automatic dotplots ("heaps"), heap + intervals, and quantile dotplots (ggplot geom)

Geoms and stats for creating dotplots ("heaps") that automatically determines a bin width that ensures the plot fits within the available space. Also ensures dots do not overlap, and allows generation of quantile dotplots using the quantiles argument to stat_heapinterval/stat_heap and stat_dist_heapinterval/stat_dist_heap. Generally follows the naming scheme and arguments of the geom_slabinterval and stat_slabinterval family of geoms and stats.

geom_heapinterval(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., dotsize = 1, stackratio = 1,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

geom_heapintervalh(..., orientation = "horizontal")

geom_heap(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_heaph(..., orientation = "horizontal")

stat_heapinterval(mapping = NULL, data = NULL, geom = "heapinterval",
  position = "identity", ..., quantiles = NA,
  point_interval = median_qi, na.rm = FALSE, show.legend = c(size =
  FALSE), inherit.aes = TRUE)

stat_heapintervalh(..., orientation = "horizontal")

stat_heap(mapping = NULL, data = NULL, geom = "heap",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)

stat_heaph(..., orientation = "horizontal")

stat_dist_heapinterval(mapping = NULL, data = NULL,
  geom = "heapinterval", position = "identity", ..., quantiles = 100,
  na.rm = FALSE, show.legend = c(size = FALSE), inherit.aes = TRUE)

stat_dist_heapintervalh(..., orientation = "horizontal")

stat_dist_heap(mapping = NULL, data = NULL, geom = "heap",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)

stat_dist_heaph(..., orientation = "horizontal")

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat
The statistical transformation to use on the data for this layer, as a string.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed to layer.

dotsize
The size of the dots relative to the bin width. The default, 1, makes dots be just about as wide as the bin width.

stackratio
The distance between the center of the dots in the same stack relative to the bin height. The default, 1, makes dots in the same stack just touch each other.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

orientation
Whether this geom is drawn horizontally ("horizontal") or vertically ("vertical"). When horizontal (resp. vertical), the geom uses the y (resp. x) aesthetic to identify different groups, then for each group uses the x (resp. y) aesthetic and the thickness aesthetic to draw a function as an slab, and draws points and intervals horizontally (resp. vertically) using the xmin, x, and xmax (resp. ymin, y, and ymax) aesthetics.

geom
Use to override the default connection between stat_slabinterval and geom_slabinterval

quantiles
For the stat_ and stat_dist_ stats, setting this to a value other than NA will produce a quantile dotplot: that is, a dotplot of quantiles from the sample (for stat_) or a dotplot of quantiles from the distribution (for stat_dist_). The value of quantiles determines the number of quantiles to plot. See Kay et al. (2016) and Fernandes et al. (2018) for more information on quantile dotplots.

point_interval
A function from the point_interval family (e.g., median_qi, mean_qi, etc). This function should take in a vector of value, and should obey the .width and .simple_names parameters of point_interval functions, such that when given a vector with .simple_names = TRUE should return a data frame with variables .value, .lower, .upper, and .width. Output will be converted to the appropriate x- or y-based aesthetics depending on the value of orientation. See the point_interval family of functions for more information.

Details

The "heap" geoms are similar to geom_dotplot but with a number of differences:

Heaps act like slabs in geom_slabinterval and can be given x positions (or y positions when in a horizontal orientation).

Given the available space to lay out dots, heaps will automatically determine how many bins to use to fit the available space.

Heaps use a center-out layout algorithm that guarantees that symmetrical data results in a symmetrical plot and avoids overlapping dots.

The shape of the dots in a heap can be changed using the slab_shape aesthetic (when using the heapinterval family) or the shape or slab_shape aesthetic (when using the heap family)

The stat_... and stat_dist_... versions of the stats when used with the quantiles argument are particularly useful for constructing quantile dotplots, which can be an effective way to communicate uncertainty using a frequency framing that may be easier for laypeople to understand (Kay et al. 2016, Fernandes et al. 2018).

Aesthetics

These stats support the following aesthetics:

x

y

datatype

thickness

size

group

In addition, in their default configuration (paired with geom_heapinterval) the following aesthetics are supported by the underlying geom:

slab_shape

datatype

alpha

colour

linetype

fill

shape

stroke

point_colour

point_fill

point_alpha

point_size

size

interval_colour

interval_alpha

interval_size

interval_linetype

slab_size

slab_colour

slab_fill

slab_alpha

slab_linetype

y

ymin

ymax

x

xmin

xmax

width

height

thickness

group

See examples of some of these aesthetics in action in vignette("slabinterval"). Learn more about the sub-geom aesthetics (like interval_color) in the scales documentation. Learn more about basic ggplot aesthetics in vignette("ggplot2-specs").

References

Kay, M., Kola, T., Hullman, J. R., & Munson, S. A. (2016). When (ish) is My Bus? User-centered Visualizations of Uncertainty in Everyday, Mobile Predictive Systems. Conference on Human Factors in Computing Systems - CHI '16, 5092--5103. doi: 10.1145/2858036.2858558 .

Fernandes, M., Walls, L., Munson, S., Hullman, J., & Kay, M. (2018). Uncertainty Displays Using Quantile Dotplots or CDFs Improve Transit Decision-Making. Conference on Human Factors in Computing Systems - CHI '18. doi: 10.1145/3173574.3173718 .

See also

Examples

# TODO