Custom ggplot scales for geom_slabinterval (and derivatives)

These scales allow more specific aesthetic mappings to be made when using geom_slabinterval and stats/geoms based on it (like eye plots).

scale_point_colour_discrete(..., aesthetics = "point_colour")

scale_point_color_discrete(..., aesthetics = "point_colour")

scale_point_colour_continuous(..., aesthetics = "point_colour",
  guide = "colourbar2")

scale_point_color_continuous(..., aesthetics = "point_colour",
  guide = "colourbar2")

scale_point_fill_continuous(..., aesthetics = "point_fill",
  guide = "colourbar2")

scale_point_size_continuous(..., range = c(1, 6))

scale_point_size_discrete(..., range = c(1, 6), na.translate = FALSE)

scale_interval_size_continuous(..., range = c(1, 6))

scale_interval_size_discrete(..., range = c(1, 6),
  na.translate = FALSE)

scale_interval_linetype(..., na.value = "blank")

scale_interval_colour_discrete(..., aesthetics = "interval_colour")

scale_interval_color_discrete(..., aesthetics = "interval_colour")

scale_interval_colour_continuous(..., aesthetics = "interval_colour",
  guide = "colourbar2")

scale_interval_color_continuous(..., aesthetics = "interval_colour",
  guide = "colourbar2")

scale_slab_colour_discrete(..., aesthetics = "slab_colour")

scale_slab_color_discrete(..., aesthetics = "slab_colour")

scale_slab_colour_continuous(..., aesthetics = "slab_colour",
  guide = "colourbar2")

scale_slab_color_continuous(..., aesthetics = "slab_colour",
  guide = "colourbar2")

scale_slab_linetype(..., na.value = "blank")

scale_slab_size_continuous(..., range = c(1, 6))

scale_slab_size_discrete(..., range = c(1, 6), na.translate = FALSE)

guide_colourbar2(...)

guide_colorbar2(...)

Arguments

...
Arguments passed to underlying scale or guide functions. E.g. scale_point_color_discrete passes arguments to scale_color_discrete. See those functions for more details.

aesthetics
Names of aesthetics to set scales for.

guide
Guide to use for legends for an aesthetic.

range
a numeric vector of length 2 that specifies the minimum and maximum size of the plotting symbol after transformation.

na.translate
In discrete scales, should we show missing values?

na.value
When na.translate is true, what value should be shown?

Details

The following additional scales / aesthetics are defined for use with geom_slabinterval and related geoms:

scale_point_color_* Point color

scale_point_fill_* Point fill color

scale_point_size_* Point size

scale_interval_color_* Interval line color

scale_interval_size_* Interval line width

scale_interval_linetype_* Interval line type

scale_slab_color_* Slab outline color

scale_slab_size_* Slab outline line width

scale_slab_linetype_* Slab outline line type

See the corresponding scale documentation in ggplot for more information; e.g. scale_color_discrete, scale_color_continuous, etc.

Other scale functions can be used with the aesthetics/scales defined here by using the aesthetics argument to that scale function. For example, to use color brewer scales with the point_color aesthetic:

scale_color_brewer(..., aesthetics = "point_color")

With continuous color scales, you may also need to provide a guide as the default guide does not work properly; this is what guide_colorbar2 is for:

scale_color_distiller(..., guide = "colorbar2", aesthetics = "point_color")

See also

Examples

# TODO