Extract samples of parameters in a Bayesian model fit into a tidy data format

Extract samples from a Bayesian/MCMC sampler for a variable with the given named indices into one of two types of long-format data frames.

spread_samples(model, ..., regex = FALSE, sep = "[, ]")

gather_samples(model, ..., regex = FALSE, sep = "[, ]")

Arguments

model
A supported Bayesian model fit / MCMC object. Currently supported models include mcmc, mcmc.list, runjags, stanfit, stanreg-objects, brm, and anything with its own as.mcmc.list implementation. If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

...
Expressions in the form of variable_name[index_1, index_2, ...] | wide_index. See `Details`.

regex
If TRUE, parameter names are treated as regular expressions and all column matching the regular expression and number of indices are included in the output. Default FALSE.

sep
Separator used to separate indices in parameter names, as a regular expression.

Value

A data frame.

Details

Imagine a JAGS or Stan fit named fit. The model may contain a parameter named b[i,v] (in the JAGS or Stan language) with i in 1:100 and v in 1:3. However, samples returned from JAGS or Stan in R will not reflect this indexing structure, instead they will have multiple columns with names like "b[1,1]", "b[2,1]", etc.

spread_samples and gather_samples provide a straightforward syntax to translate these columns back into properly-indexed variables in two different tidy data frame formats, optionally recovering index types (e.g. factor levels) as it does so.

spread_samples and gather_samples return data frames already grouped by all indices used on the variables you specify.

The difference between spread_samples is that names of parameters in the model will be spread across the data frame as column names, whereas gather_samples will gather terms into a single column named "term" and place estimates of terms into a column names "estimate". The "term" and "estimate" naming scheme is used in order to be consistent with output from the tidy function in the broom package, to make it easier to use tidybayes with broom for model comparison.

For example, spread_samples(fit, a[i], b[i,v]) might return a grouped data frame (grouped by i and v), with:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:5

column "v": value in 1:10

column "a": value of "a[i]" for iteration number ".iteration" on chain number ".chain"

column "b": value of "b[i,v]" for iteration number ".iteration" on chain number ".chain"

gather_samples(fit, a[i], b[i,v]) on the same fit would return a grouped data frame (grouped by i and v), with:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:5

column "v": value in 1:10, or NA if "term" is "a".

column "term": value in c("a", "b").

column "estimate": value of "a[i]" (when "term" is "a") or "b[i,v]" (when "term" is "b") for iteration number ".iteration" on chain number ".chain"

spread_samples and gather_samples can use type information applied to the fit object by recover_types to convert columns back into their original types. This is particularly helpful if some of the indices in your model were originally factors. For example, if the v index in the original data frame data was a factor with levels c("a","b","c"), then we could use recover_types before spread_samples:

fit %>%
 recover_types(data) 
 spread_samples(fit, b[i,v])

Which would return the same data frame as above, except the "v" column would be a value in c("a","b","c") instead of 1:3.

For variables that do not share the same subscripts (or share some but not all subscripts), we can supply their specifications separately. For example, if we have a variable d[i] with the same i subscript as b[i,v], and a variable x with no subscripts, we could do this:

spread_samples(fit, x, d[i], b[i,v])

Which is roughly equivalent to this:

spread_samples(fit, x) %>%
 inner_join(spread_samples(fit, d[i])) %>%
 inner_join(spread_samples(fit, b[i,v])) %>%
 group_by(i,v)

Similarly, this:

gather_samples(fit, x, d[i], b[i,v])

Is roughly equivalent to this:

bind_rows(
 gather_samples(fit, x),
 gather_samples(fit, d[i]),
 gather_samples(fit, b[i,v])
)

The c and cbind functions can be used to combine multiple variable names that have the same indices. For example, if we have several variables with the same subscripts i and v, we could do either of these:

spread_samples(fit, c(w, x, y, z)[i,v])

spread_samples(fit, cbind(w, x, y, z)[i,v])

# equivalent

Each of which is roughly equivalent to this:

spread_samples(fit, w[i,v], x[i,v], y[i,v], z[i,v])

Besides being more compact, the c()-style syntax is currently also faster (though that may change).

Indices can be omitted from the resulting data frame by leaving their names blank; e.g. spread_samples(fit, b[,v]) will omit the first index of b from the output. This is useful if an index is known to contain all the same value in a given model.

The shorthand .. can be used to specify one column that should be put into a wide format and whose names will be the base variable name, plus a dot ("."), plus the value of the index at ... For example:

spread_samples(fit, b[i,..]) would return a grouped data frame (grouped by i), with:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:20

column "b.1": value of "b[i,1]" for iteration number ".iteration" on chain number ".chain"

column "b.2": value of "b[i,2]" for iteration number ".iteration" on chain number ".chain"

column "b.3": value of "b[i,3]" for iteration number ".iteration" on chain number ".chain"

An optional clause in the form | wide_index can also be used to put the data frame into a wide format based on wide_index. For example, this:

spread_samples(fit, b[i,v] | v)

is roughly equivalent to this:

spread_samples(fit, b[i,v]) %>% spread(v,b)

The main difference between using the | syntax instead of the .. syntax is that the | syntax respects prototypes applied to indices with recover_types, and thus can be used to get columns with nicer names. For example:

fit %>% recover_types(data) %>% spread_samples(b[i,v] | v) would return a grouped data frame (grouped by i), with:

column ".chain": the chain number

column ".iteration": the interation number

column "i": value in 1:20

column "a": value of "b[i,1]" for iteration number ".iteration" on chain number ".chain"

column "b": value of "b[i,2]" for iteration number ".iteration" on chain number ".chain"

column "c": value of "b[i,3]" for iteration number ".iteration" on chain number ".chain"

Finally, parameter names can be regular expressions by setting regex = TRUE; e.g.:

spread_samples(fit, `b_.*`[i], regex = TRUE)

Would return a tidy data frame with parameters starting with `b_` and having one index.

See also

Examples

library(magrittr)
library(ggplot2)

data(RankCorr, package = "tidybayes")

RankCorr %>%
  spread_samples(b[i, j])

#> # A tibble: 72,000 x 5
#> # Groups:   i, j [72]
#>    .chain .iteration     i     j        b
#>     <int>      <int> <int> <int>    <dbl>
#>  1      1          1     1     1 -1.16   
#>  2      1          1     1     2 -1.69   
#>  3      1          1     1     3  0.282  
#>  4      1          1     1     4 -0.295  
#>  5      1          1     2     1 -0.587  
#>  6      1          1     2     2  0.00727
#>  7      1          1     2     3 -0.0850 
#>  8      1          1     2     4  0.154  
#>  9      1          1     3     1 -0.413  
#> 10      1          1     3     2 -0.769  
#> # ... with 71,990 more rows

RankCorr %>%
  spread_samples(b[i, j], tau[i], u_tau[i])

#> # A tibble: 72,000 x 7
#> # Groups:   i, j [72]
#>    .chain .iteration     i     j        b   tau u_tau
#>     <int>      <int> <int> <int>    <dbl> <dbl> <dbl>
#>  1      1          1     1     1 -1.16     7.13  6.65
#>  2      1          1     1     2 -1.69     7.13  6.65
#>  3      1          1     1     3  0.282    7.13  6.65
#>  4      1          1     1     4 -0.295    7.13  6.65
#>  5      1          1     2     1 -0.587    3.08  4.94
#>  6      1          1     2     2  0.00727  3.08  4.94
#>  7      1          1     2     3 -0.0850   3.08  4.94
#>  8      1          1     2     4  0.154    3.08  4.94
#>  9      1          1     3     1 -0.413    3.48  5.51
#> 10      1          1     3     2 -0.769    3.48  5.51
#> # ... with 71,990 more rows

RankCorr %>%
  gather_samples(b[i, j], tau[i], u_tau[i])

#> # A tibble: 108,000 x 6
#> # Groups:   i, j, term [108]
#>    .chain .iteration     i     j term  estimate
#>     <int>      <int> <int> <int> <chr>    <dbl>
#>  1      1          1     1     1 b     -1.16   
#>  2      1          1     1     2 b     -1.69   
#>  3      1          1     1     3 b      0.282  
#>  4      1          1     1     4 b     -0.295  
#>  5      1          1     2     1 b     -0.587  
#>  6      1          1     2     2 b      0.00727
#>  7      1          1     2     3 b     -0.0850 
#>  8      1          1     2     4 b      0.154  
#>  9      1          1     3     1 b     -0.413  
#> 10      1          1     3     2 b     -0.769  
#> # ... with 107,990 more rows

RankCorr %>%
  gather_samples(tau[i], typical_r) %>%
  mean_qi()

#> # A tibble: 19 x 6
#> # Groups:   i, term [19]
#>        i term      estimate conf.low conf.high .prob
#>    <int> <chr>        <dbl>    <dbl>     <dbl> <dbl>
#>  1     1 tau          6.03     5.07      7.14   0.95
#>  2     2 tau          3.38     2.41      4.58   0.95
#>  3     3 tau          3.65     2.68      4.86   0.95
#>  4     4 tau          6.66     5.46      8.01   0.95
#>  5     5 tau          5.13     4.25      6.03   0.95
#>  6     6 tau          6.40     5.36      7.45   0.95
#>  7     7 tau          7.89     6.63      9.28   0.95
#>  8     8 tau          6.44     5.29      7.65   0.95
#>  9     9 tau          2.68     1.84      3.72   0.95
#> 10    10 tau          6.10     4.99      7.32   0.95
#> 11    11 tau          5.93     4.93      6.99   0.95
#> 12    12 tau          7.26     6.07      8.52   0.95
#> 13    13 tau          4.35     3.62      5.14   0.95
#> 14    14 tau          3.80     2.79      4.89   0.95
#> 15    15 tau          4.19     3.46      4.95   0.95
#> 16    16 tau          3.92     2.70      5.37   0.95
#> 17    17 tau          3.22     2.67      3.84   0.95
#> 18    18 tau          3.83     2.97      4.76   0.95
#> 19    NA typical_r    0.557    0.313     0.787  0.95