Meta-stat for computing slab functions and interval functions (ggplot stat)

A meta-stat for computing slab and interval functions for use with geom_slabinterval and its derivatives. Generally speaking not intended to be used directly: The API for this stat is experimental and subject to change. This is used as the basis for several other more directly useful stats whose APIs are more stable; it is recommended to use those instead.

stat_slabinterval(mapping = NULL, data = NULL, geom = "slabinterval",
  position = "identity", ..., orientation = c("vertical",
  "horizontal"), limits_function = NULL, limits_args = list(),
  limits = NULL, slab_function = NULL, slab_args = list(), n = 501,
  interval_function = NULL, interval_args = list(),
  point_interval = NULL, .width = c(0.66, 0.95), show_slab = TRUE,
  show_interval = TRUE, na.rm = FALSE, show.legend = c(size = FALSE),
  inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

geom
Use to override the default connection between stat_slabinterval and geom_slabinterval

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed to layer. They may also be arguments to the paired geom (e.g., geom_pointinterval)

orientation
Whether this geom is drawn horizontally ("horizontal") or vertically ("vertical"). When horizontal (resp. vertical), the geom uses the y (resp. x) aesthetic to identify different groups, then for each group uses the x (resp. y) aesthetic and the thickness aesthetic to draw a function as an slab, and draws points and intervals horizontally (resp. vertically) using the xmin, x, and xmax (resp. ymin, y, and ymax) aesthetics.

limits_function
A function that takes a data frame of aesthetics and returns a data frame with columns .lower and .upper indicating the limits of the input for the slab function for that data frame.

limits_args
Additional arguments passed to limits_function

limits
Limits for slab_function, as a vector of length two. These limits are combined with those computed by the limits_function as well as the limits defined by the scales of the plot to determine the limits used to draw the slab functions: these limits specify the maximal limits; i.e., if specified, the limits will not be wider than these (but may be narrower). Use NA to leave a limit alone; e.g. limits = c(0, NA) will ensure that the lower limit does not go below 0.

slab_function
A function that takes a data frame of aesthetics and an input parameter (a vector of function inputs), and returns a data frame with columns .input (from the input vector) and .value (result of applying the function to each value of input). Given the results of slab_function, .value will be translated into the thickness aesthetic and input will be translated into either the x or y aesthetic automatically depending on the value of orientation.

slab_args
Additional arguments passed to limits_function

n
Number of points at which to evaluate slab_function

interval_function
Custom function for generating intervals (for most common use cases the point_interval argument will be easier to use). This function takes a data frame of aesthetics and a .width parameter (a vector of interval widths), and returns a data frame with columns .width (from the .width vector), .value (point summary) and .lower and .upper (endpoints of the inverals, given the .width). Output will be converted to the appropriate x- or y-based aesthetics depending on the value of orientation. If interval_function is NULL, point_interval is used instead.

interval_args
Additional arguments passed to interval_function or point_interval.

point_interval
A function from the point_interval family (e.g., median_qi, mean_qi, etc). This function should take in a vector of value, and should obey the .width and .simple_names parameters of point_interval functions, such that when given a vector with .simple_names = TRUE should return a data frame with variables .value, .lower, .upper, and .width. Output will be converted to the appropriate x- or y-based aesthetics depending on the value of orientation. See the point_interval family of functions for more information.

.width
The .width argument passed to interval_function or point_interval.

show_slab
Should the slab portion of the geom be drawn? Default TRUE.

show_interval
Should the interval portion of the geom be drawn? Default TRUE.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
Should this layer be included in the legends? Default is c(size = FALSE), unlike most geoms, to match its common use cases. FALSE hides all legends, TRUE shows all legends, and NA shows only those that are mapped (the default for most geoms).

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

See also

Examples

#TODO