Models supported by tidybayes

Tidybayes supports two classes of models and sample formats: Models/formats that provide prediction functions, and those that do not.

All Supported Models/Sample Formats

All supported models/formats support the base tidybayes sample extraction functions, such as tidy_draws, spread_draws, and gather_draws. These models/formats include:

rstan models

brm models

rstanarm models

runjags models

jags.model models, if sampled using coda.samples

jags models

MCMCglmm models

mcmc and mcmc.list objects, which are output by several model types.

Any object with an implementation of as.mcmc.list. For a list of those available in your environment, run methods(as.mcmc.list)

If you install the tidybayes.rethinking package (available at https://github.com/mjskay/tidybayes.rethinking), map and map2stan models from the rethinking package are also supported.

Models Supporting Prediction

In addition, the following models support fit and prediction extraction functions, such as add_fitted_draws and add_predicted_draws:

brm models

rstanarm models

If you install the tidybayes.rethinking package, models from the rethinking package are also supported. Note that in tidybayes.rethinking, tidy_link takes the place of add_fitted_draws and tidy_sim takes the place of add_predicted_draws.

Extending tidybayes

To include basic support for new models, one need only implement the tidy_draws generic function for that model.

To include support for estimation and prediction, one must implement the fitted_draws and predicted_draws generic functions.