Extract a tidy data frame of draws of posterior distributions of "estimated marginal means" (emmeans/lsmeans) from a Bayesian model fit.

Extract draws from the result of a call to emmeans (formerly lsmeans) or ref_grid applied to a Bayesian model.

gather_emmeans_draws(object, value = ".value")

Arguments

object
An emmGrid object such as returned by ref_grid or emmeans.

value
The name of the output column to use to contain the values of draws. Defaults to ".value".

Value

A tidy data frame of draws. The columns of the reference grid are returned as-is, with an additional column called .value (by default) containing marginal draws. The resulting data frame is grouped by the columns from the reference grid to make use of summary functions like point_interval straightforward.

Details

emmeans provides a convenient syntax for generating draws from "estimated marginal means" from a model, and can be applied to various Bayesian models, like stanreg-objects and MCMCglmm. Given a ref_grid object as returned by functions like ref_grid or emmeans applied to a Bayesian model, gather_emmeans_draws returns a tidy format data frame of draws from the marginal posterior distributions generated by emmeans.

See also

Examples

library(dplyr)
library(magrittr)
library(rstanarm)
library(emmeans)

#> NOTE: As of emmeans versions > 1.2.3,
#>       The 'cld' function will be deprecated in favor of 'CLD'.
#>       You may use 'cld' only if you have package:multcomp attached.

# Here's an example dataset with a categorical predictor (`condition`) with several levels:
set.seed(5)
n = 10
n_condition = 5
ABC =
  data_frame(
    condition = rep(c("A","B","C","D","E"), n),
    response = rnorm(n * 5, c(0,1,2,1,-1), 0.5)
  )

m = stan_glm(response ~ condition, data = ABC,
  # 1 chain / few iterations just so example runs quickly
  # do not use in practice
  chains = 1, iter = 500)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 0 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 500 [  0%]  (Warmup)
#> Iteration:  50 / 500 [ 10%]  (Warmup)
#> Iteration: 100 / 500 [ 20%]  (Warmup)
#> Iteration: 150 / 500 [ 30%]  (Warmup)
#> Iteration: 200 / 500 [ 40%]  (Warmup)
#> Iteration: 250 / 500 [ 50%]  (Warmup)
#> Iteration: 251 / 500 [ 50%]  (Sampling)
#> Iteration: 300 / 500 [ 60%]  (Sampling)
#> Iteration: 350 / 500 [ 70%]  (Sampling)
#> Iteration: 400 / 500 [ 80%]  (Sampling)
#> Iteration: 450 / 500 [ 90%]  (Sampling)
#> Iteration: 500 / 500 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.017 seconds (Warm-up)
#>                0.016 seconds (Sampling)
#>                0.033 seconds (Total)
#> 

# Once we've fit the model, we can use emmeans() (and functions
# from that package) to get whatever marginal distributions we want.
# For example, we can get marginal means by condition:
m %>%
  emmeans(~ condition) %>%
  gather_emmeans_draws() %>%
  median_qi()

#> # A tibble: 5 x 7
#>   condition .value  .lower .upper .width .point .interval
#>   <fct>      <dbl>   <dbl>  <dbl>  <dbl> <chr>  <chr>    
#> 1 A          0.189 -0.0845  0.553   0.95 median qi       
#> 2 B          1.00   0.636   1.37    0.95 median qi       
#> 3 C          1.85   1.51    2.17    0.95 median qi       
#> 4 D          1.02   0.703   1.40    0.95 median qi       
#> 5 E         -0.927 -1.23   -0.634   0.95 median qi       

# or we could get pairwise differences:
m %>%
  emmeans(~ condition) %>%
  contrast(method = "pairwise") %>%
  gather_emmeans_draws() %>%
  median_qi()

#> # A tibble: 10 x 7
#>    contrast   .value .lower .upper .width .point .interval
#>    <fct>       <dbl>  <dbl>  <dbl>  <dbl> <chr>  <chr>    
#>  1 A - B    -0.810   -1.31  -0.305   0.95 median qi       
#>  2 A - C    -1.65    -2.10  -1.15    0.95 median qi       
#>  3 A - D    -0.828   -1.27  -0.328   0.95 median qi       
#>  4 A - E     1.12     0.665  1.58    0.95 median qi       
#>  5 B - C    -0.840   -1.33  -0.326   0.95 median qi       
#>  6 B - D    -0.00630 -0.504  0.494   0.95 median qi       
#>  7 B - E     1.94     1.49   2.39    0.95 median qi       
#>  8 C - D     0.838    0.278  1.30    0.95 median qi       
#>  9 C - E     2.81     2.33   3.21    0.95 median qi       
#> 10 D - E     1.96     1.52   2.40    0.95 median qi       

# see the documentation of emmeans() for more examples of types of
# contrasts supported by that packge.